chore(publish): utilities for creating a webview panel
[Docs](https://github.com/dendronhq/dendron-docs/blob/main/vault/pkg.dendron-plugin-views.dev.cook.md#L103:L103)
